refactor(ivy): move `styling/interfaces.ts` to `interfaces/styling.ts` (#32731)
PR Close #32731
