+++ packages/core/src/debug/debug_node.ts
@@ -10,9 +10,9 @@ import {Injector} from '../di';
 import {getViewComponent} from '../render3/global_utils_api';
 import {CONTAINER_HEADER_OFFSET, LContainer, NATIVE} from '../render3/interfaces/container';
 import {TElementNode, TNode, TNodeFlags, TNodeType} from '../render3/interfaces/node';
+import {StylingMapArray, TStylingContext} from '../render3/interfaces/styling';
 import {isComponentHost, isLContainer} from '../render3/interfaces/type_checks';
 import {LView, PARENT, TData, TVIEW, T_HOST} from '../render3/interfaces/view';
-import {StylingMapArray, TStylingContext} from '../render3/styling/interfaces';
 import {NodeStylingDebug} from '../render3/styling/styling_debug';
 import {isStylingContext, stylingMapToStringMap} from '../render3/styling/util';
 import {getComponent, getContext, getInjectionTokens, getInjector, getListeners, getLocalRefs, isBrowserEvents, loadLContext} from '../render3/util/discovery_utils';
