+++ packages/core/src/render3/instructions/lview_debug.ts
@@ -18,8 +18,8 @@ import {PropertyAliases, TContainerNode, TElementNode, TNode as ITNode, TNode, T
 import {SelectorFlags} from '../interfaces/projection';
 import {TQueries} from '../interfaces/query';
 import {RComment, RElement, RNode} from '../interfaces/renderer';
+import {TStylingContext} from '../interfaces/styling';
 import {BINDING_INDEX, CHILD_HEAD, CHILD_TAIL, CLEANUP, CONTEXT, DECLARATION_VIEW, ExpandoInstructions, FLAGS, HEADER_OFFSET, HOST, HookData, INJECTOR, LView, LViewFlags, NEXT, PARENT, QUERIES, RENDERER, RENDERER_FACTORY, SANITIZER, TData, TVIEW, TView as ITView, TView, T_HOST} from '../interfaces/view';
-import {TStylingContext} from '../styling/interfaces';
 import {DebugStyling as DebugNewStyling, NodeStylingDebug} from '../styling/styling_debug';
 import {isStylingContext} from '../styling/util';
 import {attachDebugObject} from '../util/debug_utils';
